// pronti-frontend-main/front/src/app/modulos/cuota/servicios/dto/cuota.update-dto.ts
export interface CuotaUpdateDto {

  // nombre?: string;

  fechaCobro?: string;
  fechaMora?: string;
  numeroCuota?: number;
  valorCuota?: number;
  valorPagadoCuota?: number;
  valorTasaAdministrativa?: number;
  valorImpuesto?: number;
  abonoCapital?: number;
  estaPagado?: string;
  estaMora?: string;

}